refactor(effects): MysticBackground 순수 함수 mysticUtils.ts로 분리
